# bruno/openapi.yml
openapi: 3.0.3
info:
  title: VizHome API
  version: 1.0.0
  description: API REST de la plateforme VizHome — visualisation 3D architecturale
    propulsée par IA (Gemini).
  contact:
    email: support@vizhome.fr
  license:
    name: Propriétaire
paths:
...
  /api/v1/me/:
    get:
      operationId: me_retrieve
      description: GET = profil complet (user + stats + preferences), PATCH = update
        profil.
      tags:
      - me
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: ''
    put:
      operationId: me_update
      description: GET = profil complet (user + stats + preferences), PATCH = update
        profil.
      tags:
      - me
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UserRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UserRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: ''
    patch:
      operationId: me_partial_update
      description: GET = profil complet (user + stats + preferences), PATCH = update
        profil.
      tags:
      - me
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedUpdateMeRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedUpdateMeRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedUpdateMeRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateMe'
          description: ''
...
components:
  schemas:
...
    FontSizeEnum:
      enum:
      - small
      - medium
      - large
      type: string
      description: |-
        * `small` - Petit
        * `medium` - Moyen
        * `large` - Grand
...
    LanguageEnum:
      enum:
      - fr
      - en
      - es
      - de
      type: string
      description: |-
        * `fr` - Français
        * `en` - English
        * `es` - Español
        * `de` - Deutsch
...
    PatchedUpdateMeRequest:
      type: object
      description: Subset des champs modifiables via PATCH /me.
      properties:
        first_name:
          type: string
          maxLength: 100
        last_name:
          type: string
          maxLength: 100
        avatar_url:
          type: string
          format: uri
          maxLength: 500
...
    PlanEnum:
      enum:
      - free
      - pro
      - enterprise
      type: string
      description: |-
        * `free` - Gratuit
        * `pro` - Pro
        * `enterprise` - Entreprise
...
    RenderFormatEnum:
      enum:
      - png
      - jpg
      - webp
      type: string
      description: |-
        * `png` - PNG
        * `jpg` - JPG
        * `webp` - WebP
    RenderQualityEnum:
      enum:
      - draft
      - standard
      - high
      type: string
      description: |-
        * `draft` - Brouillon
        * `standard` - Standard
        * `high` - Haute
    RenderResolutionEnum:
      enum:
      - '1024'
      - '2048'
      - '4096'
      type: string
      description: |-
        * `1024` - 1024px
        * `2048` - 2048px
        * `4096` - 4096px
...
    ThemeEnum:
      enum:
      - light
      - dark
      - system
      type: string
      description: |-
        * `light` - Clair
        * `dark` - Sombre
        * `system` - Système
...
    UpdateMe:
      type: object
      description: Subset des champs modifiables via PATCH /me.
      properties:
        first_name:
          type: string
          maxLength: 100
        last_name:
          type: string
          maxLength: 100
        avatar_url:
          type: string
          format: uri
          maxLength: 500
    User:
      type: object
      description: Serializer principal pour /me — inclut stats + preferences en nested.
      properties:
        id:
          type: integer
          readOnly: true
        email:
          type: string
          format: email
          readOnly: true
        first_name:
          type: string
          maxLength: 100
        last_name:
          type: string
          maxLength: 100
        name:
          type: string
          readOnly: true
        avatar_url:
          type: string
          format: uri
          maxLength: 500
        plan:
          allOf:
          - $ref: '#/components/schemas/PlanEnum'
          readOnly: true
        date_joined:
          type: string
          format: date-time
          readOnly: true
        stats:
          allOf:
          - $ref: '#/components/schemas/UserStats'
          readOnly: true
        preferences:
          allOf:
          - $ref: '#/components/schemas/UserPreferences'
          readOnly: true
      required:
      - date_joined
      - email
      - id
      - name
      - plan
      - preferences
      - stats
    UserPreferences:
      type: object
      properties:
        theme:
          $ref: '#/components/schemas/ThemeEnum'
        language:
          $ref: '#/components/schemas/LanguageEnum'
        notif_email_render:
          type: boolean
        notif_email_newsletter:
          type: boolean
        notif_push_render:
          type: boolean
        notif_push_mentions:
          type: boolean
        render_quality:
          $ref: '#/components/schemas/RenderQualityEnum'
        render_format:
          $ref: '#/components/schemas/RenderFormatEnum'
        render_resolution:
          $ref: '#/components/schemas/RenderResolutionEnum'
        analytics_enabled:
          type: boolean
        marketing_enabled:
          type: boolean
        two_factor_enabled:
          type: boolean
        reduced_motion:
          type: boolean
        high_contrast:
          type: boolean
        font_size:
          $ref: '#/components/schemas/FontSizeEnum'
...
    UserRequest:
      type: object
      description: Serializer principal pour /me — inclut stats + preferences en nested.
      properties:
        first_name:
          type: string
          maxLength: 100
        last_name:
          type: string
          maxLength: 100
        avatar_url:
          type: string
          format: uri
          maxLength: 500
...
    UserStats:
      type: object
      properties:
        renders_this_month:
          type: integer
          readOnly: true
        renders_limit:
          type: integer
          readOnly: true
        total_projects:
          type: integer
          readOnly: true
        storage_used_bytes:
          type: integer
          readOnly: true
        storage_limit_bytes:
          type: integer
          readOnly: true
        period_started_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - period_started_at
      - renders_limit
      - renders_this_month
      - storage_limit_bytes
      - storage_used_bytes
      - total_projects
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...